+++ src/OpenGL/libGLESv2/Program.cpp
@@ -866,7 +866,7 @@ namespace es2
 		unsigned int count = UniformComponentCount(targetUniform->type);
 
 		// Sized query - ensure the provided buffer is large enough
-		if(bufSize && *bufSize < count * sizeof(GLfloat))
+		if(bufSize && static_cast<unsigned int>(*bufSize) < count * sizeof(GLfloat))
 		{
 			return false;
 		}
@@ -925,7 +925,7 @@ namespace es2
 		unsigned int count = UniformComponentCount(targetUniform->type);
 
 		// Sized query - ensure the provided buffer is large enough
-		if(bufSize && *bufSize < count * sizeof(GLint))
+		if(bufSize && static_cast<unsigned int>(*bufSize) < count * sizeof(GLint))
 		{
 			return false;
 		}
@@ -974,7 +974,7 @@ namespace es2
 		unsigned int count = UniformComponentCount(targetUniform->type);
 
 		// Sized query - ensure the provided buffer is large enough
-		if(bufSize && *bufSize < count * sizeof(GLuint))
+		if(bufSize && static_cast<unsigned int>(*bufSize) < count * sizeof(GLuint))
 		{
 			return false;
 		}
@@ -1370,7 +1370,7 @@ namespace es2
 
 				index++;
 			}
-			while(index < registerIndex + arraySize);
+			while(index < registerIndex + static_cast<int>(arraySize));
 	    }
 
 		Uniform *uniform = 0;
@@ -1417,7 +1417,7 @@ namespace es2
 			uniforms.push_back(uniform);
 			unsigned int index = uniforms.size() - 1;
 
-			for(unsigned int i = 0; i < uniform->size(); i++)
+			for(int i = 0; i < uniform->size(); i++)
 			{
 				uniformIndex.push_back(UniformLocation(name, i, index));
 			}
